Vray for Maya Domemaster3D Update
This update adds significant Vray for Maya Domemaster3D changes that
allow the DomemasterStereo and LatLongStereo shaders to work in the Maya
Render View and the Vray Frame Buffer window.

This is done through the use of a Render Settings **Mel/Python
Callbacks** `Post Translate Python Script` item that is added
automatically when a DomemasterStereo or LatLongStereo Vray Extra
Attribute is attached to a camera Shape node and enabled with the
checkbox.

Note: This is an experimental Vray for Maya update that is being posted
so the other Domemaster3D developers can try it out and make comments.
These Vray additions have not had a lot of testing done.