Domemaster and Latlong for 3ds Max\Vray
Working, fully featured, versions of the Domemaster and Latlong cameras
for Vray in 3ds Max (tested on Max 2015)

Still to finalize:

Use of Texmaps instead of Bitmaps for the camera separation multiplier,
turn multiplier, and head tilt.

User Interface, currently pretty rough.
Once a bitmap is assigned button do not refresh automatically (switch to
another Command panel tab and back to see the bitmap name in the
button).
Bitmap buttons not labeled. (Top is camera separation multiplier, below
is head turn multipliar, lower is head tilt)

The Latlong can use a negative angle to restrict the rendering to the
top part of the dome.