Version 1.0 of Vray Domemaster and latlong cameras for 3ds max
This is the fully featured and functional :) version 1.0 of the Vray
Domemaster and Latlong stereo cameras source code.

If you used the beta versions this release will break compatibility
since the paramblock 2 is changed to add some features.
The old cameras won't load, and will need to be re-created.

Changes for both cameras

1) New Class ID for a fresh start.
Paramblock2 is changed anyway and old cameras woudln't load.

2) Neck Offset
This new parameter moves the cameras the specified offset along the view
direction.
It simulates the fact that most real camera rigs have the cameras
mounted at a forward offset from the actual pivot point. If you need to
match a live plate this parameter will help avoid shifting between
plates and CG elements.

3) Map buttons now update the texture name after the selection is done.

4) Minor cosmetic UI chages

LatLong only changes

4) Exposed a new option "Hemirect."
When this checkbox is selected the Vertical FOV is now interpreted as
half the zenith FOV.
An FOV of 45 will render 45 degrees from the Zenith (90 degrees
effective FOV)
This was previously an hidden feature exposed by using a negative FOV
value.

5) Flip Y fixed to correctly mirror the image when Hemirect is selected.
Previously this was not working when using the hidden negative vertical
FOV described above.

Still to do is the use of Texmaps instead of Bitmaps for the maps.