Vray for 3DS Max Shaders Added
This update adds an initial build of the Domemaster3D shaders that work
with Vray 3.0 for 3DS Max 2015, Vray 3.0 for 3DS Max 2014, and Vray 2.5
for 3DS Max 2015. The new vray shaders are beta versions that are being
posted so the beta testers can evaluate them. The shaders have been
compiled with Visual Studios 2013 so you may have to download the
Microsoft Visual Studio 2013 x64 redistributable library to use them.