Softimage Arnold and Vray Shader Update
Updated the Domemaster3D for SItoA SPDL files to match the MtoA gui
controls. Added a new set of Vray for Softimage SPDL files as part of a
work in progress task of adding Domemaster3D shader support. The Vray
for Softimage compiled DomemasterStereo and LatLongStereo shaders still
need to have the Vray Translator figured out so the lens shaders can be
activated in the Vray Frame Buffer and the .vrscene exporter. Updated
the Mental Ray Domemaster3D for Maya menus to separate the Maya 2010 XPM
icons for the shelf, from Maya 2011 to 2015 PNG shelf icons.