Arnold Domemaster3D v1.7 Update
Updated the Maya shelf tools to add a new Automagic LatLong Stereo
option

Updated the python script that creates the Arnold LatLongStereo and
DomeStereo cameras. The stereo rig manager defaultRig value is now
switched automatically when a LatLongStereo or DomeStereo camera is
created. This fixes an alignment issue that happened when the wrong
stereo camera defaultRig was active when creating a new camera in
Arnold.

Updated the dome grid shape to use a full sphere as the default FOV
value for the wireframe mesh geometry. This improves the rendering
experience when a latlong shader is used.