Maya Shelf Tool Added - Launch a 2nd Maya Instance
Added a "MAYA" shelf tool and the "Launch a 2nd Maya Instance"
Domemaster3D menu item will launch a 2nd copy of Maya that will run
independently. This is handy if you want to work in two separate Maya
sessions at the same time. The MEL command used by the shelf tool is:

    source "domeDiagnostics.mel"; domeLaunchMayaInstance();