Added ray differential based sampling to support mr 3.14
This is the initial commit of a new domeAFL_2017 based update to the
mental ray lens shaders that adds ray differential based texture
sampling to support mental ray 3.14 which is a feature available in the
latest mental ray Beta 7 release. This change is based upon the code
suggestions from NVIDIA ARC's Enzo Catalano. Thanks!