DomeViewer Image Sequence Fix
Merry Christmas! 🎉 This update improves the stereoscopic image
sequence playback support in the DomeViewer tool.

The code revision uses the showEditorExact() MEL function to display
the newly created file texture node in the Attribute Editor window. By
accessing the texture map once with the showEditorExact() command  it
causes Maya to start evaluating the numbered image sequence component
part of the image filename.

This code snippet is activated with an evalDeferred() call so it
happens once the DomeViewer script has completed all of its tasks and
Maya is sitting idle.